Create editable and non-editable MemoryActivity (NOT TESTED)
